+++ src/main/java/com/example/trendsetter/Controller/HinhAnhController.java
@@ -1,13 +1,19 @@
 package com.example.trendsetter.Controller;
 
+import com.example.trendsetter.Entity.ChatLieu;
 import com.example.trendsetter.Entity.HinhAnh;
+import com.example.trendsetter.Entity.SanPham;
 import com.example.trendsetter.Repository.HinhAnhRepository;
+import com.example.trendsetter.Repository.SanPhamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -24,17 +30,22 @@ public class HinhAnhController {
     @Autowired
     HinhAnhRepository hinhAnhRepository;
 
+    @Autowired
+    SanPhamRepository sanPhamRepository;
+
     @GetMapping("/hien-thi")
-    public String hienThi(Model model){
-        model.addAttribute("danhSach",hinhAnhRepository.findAll());
+    public String hienThi(@RequestParam(defaultValue = "0") int page, Model model) {
+        Page<HinhAnh> hinhAnhPage = hinhAnhRepository.findAll(PageRequest.of(page, 5));
+        model.addAttribute("danhSach",hinhAnhPage.getContent());
+        model.addAttribute("pageNumber", page);
+        model.addAttribute("totalPages", hinhAnhPage.getTotalPages());
         return "HinhAnh/hien-thi";
     }
     @PostMapping("/add")
     public String add(
         @RequestParam("urlHinhAnh") MultipartFile hinhAnhFile,
-        Model model){
+        RedirectAttributes redirectAttributes){
             HinhAnh hinhAnh = new HinhAnh();
-            hinhAnh.setTrangThai("Đang Hoạt Động");
             try {
                 // Xử lý upload hình ảnh
                 if (hinhAnhFile != null && !hinhAnhFile.isEmpty()) {
@@ -50,10 +61,11 @@ public class HinhAnhController {
                 }
 
                 // Lưu sản phẩm chi tiết
+                redirectAttributes.addFlashAttribute("successMessage", "Hình ảnh đã được thêm thành công!");
                 hinhAnhRepository.save(hinhAnh);
             } catch (IOException e) {
                 e.printStackTrace();
-                model.addAttribute("error", "Lỗi khi tải lên hình ảnh");
+                redirectAttributes.addFlashAttribute("errorMessage", "Lỗi khi tải lên hình ảnh");
                 return "HinhAnh/hien-thi";
             }
             return "redirect:/hinh-anh/hien-thi";
@@ -64,12 +76,12 @@ public class HinhAnhController {
         model.addAttribute("hinhAnh",hinhAnhRepository.findById(id));
         return "HinhAnh/hien-thi";
     }
+
     @PostMapping("/update")
     public String update(@RequestParam("id") Integer id,
                          @RequestParam("urlHinhAnh") MultipartFile hinhAnhFile,
-                         Model model){
+                         RedirectAttributes redirectAttributes){
         HinhAnh hinhAnh = hinhAnhRepository.findById(id).get();
-        hinhAnh.setTrangThai("Đang Hoạt Động");
         try {
             // Xử lý upload hình ảnh
             if (hinhAnhFile != null && !hinhAnhFile.isEmpty()) {
@@ -84,19 +96,24 @@ public class HinhAnhController {
                 hinhAnh.setUrlHinhAnh(fileName);
             }
 
-            // Lưu sản phẩm chi tiết
+            redirectAttributes.addFlashAttribute("successMessage", "Hình ảnh đã được sửa thành công!");
             hinhAnhRepository.save(hinhAnh);
         } catch (IOException e) {
             e.printStackTrace();
-            model.addAttribute("error", "Lỗi khi tải lên hình ảnh");
+            redirectAttributes.addFlashAttribute("errorMessage", "Lỗi khi tải lên hình ảnh");
             return "HinhAnh/hien-thi";
         }
         return "redirect:/hinh-anh/hien-thi";
     }
 
     @GetMapping("/delete")
-    public String delete(@RequestParam("id")Integer id){
-        hinhAnhRepository.deleteById(id);
+    public String delete(@RequestParam("id")Integer id,RedirectAttributes redirectAttributes){
+        if (hinhAnhRepository.existsById(id)){
+            hinhAnhRepository.deleteById(id);
+            redirectAttributes.addFlashAttribute("successMessage", "Hình ảnh đã được xóa thành công!");
+        }else {
+            redirectAttributes.addFlashAttribute("errorMessage", "Chất liệu không tồn tại!");
+        }
         return "redirect:/hinh-anh/hien-thi";
     }
 }
